Migrate admin routes to react-router framework modules
