+++ app/routes/admin_.email.restricted/route.tsx
