+++ app/routes/admin_.email.restricted_.$restrictedMailId/route.tsx
